// AssessmentTask_ECommerce/E-Commerce-frontend/src/app/models/product.ts
// src/app/product.ts
// export interface Product {
//     productID: number;
//     productName: string;
//     price: number;
//   }
  
  // src/app/models/product.ts
  export interface Product {
    productID: number;
    productName: string;
    price: number;
    description: string;  // Add the description field here
  }
  
  